refactor: split group-table.tsx into folder with sub-components
Closes #360
